requirements.txt
python-dotenv==0.19.2
python-telegram-bot==13.10